Turn the repo into an installable python package
As per the python package requirements, this is mostly just adding some
additional files to tell setuptools what to do, but I had to move the
code into a sub-folder.

Note I've removed requirements.txt in favour of install_requires in
setup.py and I've kept all the packages the same except one, MITIE,
because it's now available on pypi and this makes installing simpler.